Fish (#9)
* This is update 1 I changed the scanner to my own scanner. I added new 
operators and control declare. Next update will add boolean constants
ad the control flows

* This is update 2 to the scanner. It now detects the new operators, control
tokens (flow and end), declaration tokens, and booleans. Next iteration
will add current token and next token, skipping comments, and combining 
operators.

* THis is a minor update that fixed some things I noticed I forgot to do
before the last push. Added f to Booleans and replaced a few lines of
code with a function call.

* This update added comment skipping and handles getting two tokens. All
that should be left for the scanner is to handle double operators.

* With this update the scanner should do everything needed. The only thing
left in the scanner requirements is to convert the strings. I need to
update the comments, but wanted to get the code on the repository so we 
can test the rest of it tomorrow.

* Added comments to the code. All that's left is to convert the strings