Fishp3 (#12)
* Scanner now updates the iColPos and the iSourcLineNumber of the token
object.

* Added the parser class. Currently the parser detects a Declaration 
statement and adds the identifier to the symbol table if it does not 
exist already.

* Added the parser class. Currently the parser detects a Declaration 
statement and adds the identifier to the symbol table if it does not 
exist already.Updated some mistakes made.

* Updated the parser a bit. THe declaration statement now checks
for a complete statement before moving on. I also added a placeholder for 
when we add in Int i = 2; functionality.