Fishp3 (#14)
* Scanner now updates the iColPos and the iSourcLineNumber of the token
object.

* Added the parser class. Currently the parser detects a Declaration 
statement and adds the identifier to the symbol table if it does not 
exist already.

* Added the parser class. Currently the parser detects a Declaration 
statement and adds the identifier to the symbol table if it does not 
exist already.Updated some mistakes made.

* Updated the parser a bit. THe declaration statement now checks
for a complete statement before moving on. I also added a placeholder for 
when we add in Int i = 2; functionality.

* Added some placeholder code from clark's notes to the parser for later
use. Need the memory manager, Utility class, and ResultValue class. Have
To figure out what the last one is.

* Scanner - added array lists of tokens and lines. Will overhaul how scanner
          works for later implementation of loops
Parser  - Now handles print statements. Simple expression is almost done.
          Still needs the memory manager and some utility functions to
          do any actual work.

* Scanner - Scanner now runs off the array list. There is now a function called
          setLine that takes an int and loads that line for running again.
          An example of this is in the code for the while loop.
Parser  - Simple expression is mostly done, only a few more 
          possibilities to add. If else statements and while loop are added.
          once the numeric functions are added, I'll update everything to work
          with them.
Added some other files to the branch that were not there before - SMValue, StorageManager and the updated SymbolTable.